Move CGoalDrawConfiguration traits
